+++ processing/topics/parking/obstacles/0_points_project_to_kerb.sql
@@ -1,4 +1,4 @@
-DO $$ BEGIN RAISE NOTICE 'START projecting obstacle points at %', clock_timestamp(); END $$;
+DO $$ BEGIN RAISE NOTICE 'START projecting obstacle points at %', clock_timestamp() AT TIME ZONE 'Europe/Berlin'; END $$;
 
 DROP TABLE IF EXISTS _parking_obstacle_points_projected CASCADE;
 
@@ -8,7 +8,7 @@ SELECT
   p.osm_type,
   p.osm_id,
   p.id as source_id,
-  p.tags,
+  p.tags || jsonb_build_object('tag_sources', p.id, 'geom_sources', pk.kerb_id) as tags,
   p.meta,
   pk.* INTO _parking_obstacle_points_projected
 FROM