+++ processing/topics/parking/obstacles/point/obstacle_point_categories.lua
@@ -52,26 +52,6 @@ obstacle_point_categories = {
     tags = function(tags) return {} end,
     tags_cc = { 'traffic_sign', 'highway' },
   }),
-  -- TEMPORARY DISABLED
-  --
-  -- class_obstacle_category.new({
-  --   id = 'bus_stop_conditional', -- https://overpass-turbo.eu/s/25Z1
-  --   buffer_radius = function(tags) return 15 end,
-  --   conditions = function(tags)
-  --     return tags['highway'] == 'bus_stop' and tags.opening_hours ~= nil
-  --   end,
-  --   tags = function(tags) return { } end,
-  --   tags_cc = { 'name', 'note', 'ref', 'opening_hours' },
-  -- }),
-  class_obstacle_category.new({
-    id = 'bus_stop', -- https://wiki.openstreetmap.org/wiki/DE:Tag:highway%3Dbus_stop
-    buffer_radius = function(tags) return 15 end,
-    conditions = function(tags)
-      return tags['highway'] == 'bus_stop' and tags.opening_hours == nil
-    end,
-    tags = function(tags) return { } end,
-    tags_cc = { 'name', 'note', 'ref' },
-  }),
   class_obstacle_category.new({
     id = 'loading_ramp',
     buffer_radius = function(tags) return 2 end,