+++ processing/topics/parking/parking.sql
@@ -14,6 +14,7 @@
 -- CREATE CUSTOM FUNCTIONS
 -- * … TO PROJECT
 \i '/processing/topics/parking/custom_functions/project_to_k_closest_kerbs.sql'
+\i '/processing/topics/parking/custom_functions/project_to_closest_platform.sql'
 \i '/processing/topics/parking/custom_functions/project_to_line.sql'
 -- * … FOR KERBS
 \i '/processing/topics/parking/custom_functions/line_azimuth_at_index.sql'
@@ -60,6 +61,9 @@
 -- HANDLE UNPROJECTED OBSTACLES
 \i '/processing/topics/parking/obstacles_unprojected/0_filter_turning_circles.sql'
 
+-- HANDLE PUBLIC TRANSPORT
+\i '/processing/topics/parking/public_transport/0_points_project_to_kerb_and_platform.sql'
+
 -- HANDLE SEPARATE PARKINGS
 \i '/processing/topics/parking/separate_parkings/0_areas_project_to_kerb.sql'
 \i '/processing/topics/parking/separate_parkings/0_points_project_to_kerb.sql'
@@ -74,9 +78,7 @@
 \i '/processing/topics/parking/5_estimate_parking_capacities.sql'
 \i '/processing/topics/parking/6_filter_parkings.sql'
 \i '/processing/topics/parking/7_finalize_parkings.sql'
--- \i '/processing/topics/parking/8_qa_parkings_euvm_voronoi.sql'
+\i '/processing/topics/parking/8_qa_parkings_euvm_voronoi.sql'
 \i '/processing/topics/parking/9_create_labels.sql'
 
-
-
-DO $$ BEGIN RAISE NOTICE 'FINISH topics/parking/parking.sql at %', clock_timestamp(); END $$;
+DO $$ BEGIN RAISE NOTICE 'FINISH topics/parking/parking.sql at %', clock_timestamp() AT TIME ZONE 'Europe/Berlin'; END $$;