+++ processing/topics/parking/roads/0_create_kerbs.sql
@@ -1,4 +1,4 @@
-DO $$ BEGIN RAISE NOTICE 'START creating kerbs %', clock_timestamp(); END $$;
+DO $$ BEGIN RAISE NOTICE 'START creating kerbs %', clock_timestamp() AT TIME ZONE 'Europe/Berlin'; END $$;
 
 DROP TABLE IF EXISTS _parking_kerbs;