+++ processing/topics/parking/roads/1_find_intersections.sql
@@ -1,4 +1,4 @@
-DO $$ BEGIN RAISE NOTICE 'START finding intersections at %', clock_timestamp(); END $$;
+DO $$ BEGIN RAISE NOTICE 'START finding intersections at %', clock_timestamp() AT TIME ZONE 'Europe/Berlin'; END $$;
 
 -- PREPARE
 DROP TABLE IF EXISTS _parking_intersections;