+++ processing/topics/parking/roads/4_build_graph.sql
@@ -1,4 +1,4 @@
-DO $$ BEGIN RAISE NOTICE 'START building graph at %', clock_timestamp(); END $$;
+DO $$ BEGIN RAISE NOTICE 'START building graph at %', clock_timestamp() AT TIME ZONE 'Europe/Berlin'; END $$;
 
 DROP TABLE IF EXISTS _parking_edges;