+++ processing/topics/roads_bikelanes/bikelanes/BikelaneCategories.lua
@@ -5,6 +5,7 @@ require("CreateSubcategoriesAdjoiningOrIsolated")
 require("SanitizeTrafficSign")
 require("DeriveSmoothness")
 require("HighwayClasses")
+local has_tag_with_prefix = require('has_tag_with_prefix')
 local SANITIZE_ROAD_TAGS = require('sanitize_road_tags')
 local inspect = require("inspect")
 local to_semicolon_list = require('to_semicolon_list')
@@ -316,7 +317,8 @@ local footwayBicycleYes = BikelaneCategory.new({
     end
 
     -- 2. Check bicycle access: has to have bicycle=yes or the right traffic sign
-    if tags.bicycle ~= "yes" and not ContainsSubstring(SanitizeTrafficSign(tags.traffic_sign), "1022-10") then
+    local has_bicycle_access = tags.bicycle == 'yes' or ContainsSubstring(SanitizeTrafficSign(tags.traffic_sign), '1022-10')
+    if not has_bicycle_access then
       return
     end
 
@@ -412,7 +414,7 @@ local crossing = BikelaneCategory.new({
   infrastructureExists = true,
   implicitOneWay = true, -- 'oneway=implicit_yet' but actually unknown so lets be cautions
   implicitOneWayConfidence = 'low',
-  copySurfaceSmoothnessFromParent = false,
+  copySurfaceSmoothnessFromParent = true,
   condition = function(tags)
     if tags.highway == "cycleway" and tags.cycleway == "crossing" then
       return true
@@ -734,6 +736,23 @@ local needsClarification = BikelaneCategory.new({
     end
 
     if tags.highway == "path" and tags.bicycle == "designated" then
+      -- Handle exclusive cycleways that are misstagged as path.
+      -- But exclude those MTB path that are not relevant for a everyday cycle network.
+      -- NOTE: We could allow some cases later (if needed) when certain traffic_sigsn are present.
+      -- See https://github.com/FixMyBerlin/radinfra.de/issues/14
+      --
+      -- REMINDER: Also update `BikelaneTodos.lua` > `unexpected_highway_path`
+      if tags.foot == 'no' then
+        local excluded_surfaces = Set({ 'ground', 'dirt', 'fine_gravel', 'gravel', 'pebblestone', 'earth' })
+        if has_tag_with_prefix(tags, 'mtb:') or
+          tags.mtb == 'yes' or
+          excluded_surfaces[tags.surface]
+        then
+          return false
+        end
+
+        return true
+      end
       return true
     end